Split tunnel-specific crypto from aes.h/cpp.
